Refactor desktop codebase
