+++ desktop/src/main/base/ipcRouter.ts
@@ -1,0 +1,5 @@
+import { initTRPC } from "@trpc/server";
+
+const t = initTRPC.create();
+export const r = t.router;
+export const proc = t.procedure;
