+++ desktop/src/main/base/logging.ts
