+++ desktop/src/main/base/safeStorage.ts
