+++ desktop/src/main/base/settings.ts
@@ -3,7 +3,7 @@ import { safeStorage } from "./safeStorage";
 import { z } from "zod";
 import * as fsp from "fs/promises";
 import { AssetTypeSchema } from "@bowser/prisma/types";
-import { IPCEvents } from "./ipcEventBus";
+import { IPCEvents } from "../ipcEventBus";
 import { ipcMain } from "electron";
 import logging from "./logging";
