+++ desktop/src/renderer/screens/vMix.tsx
@@ -10,7 +10,7 @@ import {
 } from "@bowser/prisma/utilityTypes";
 import { z } from "zod";
 import { VMIX_NAMES } from "../../common/constants";
-import { ListInput } from "../../main/vmixTypes";
+import { ListInput } from "../../main/vmix/vmixTypes";
 import invariant from "../../common/invariant";
 import { Alert } from "@bowser/components/alert";
 import { Progress } from "@bowser/components/progress";
